+++ src/main/java/kattsyn/dev/ApartmentsUnderConstruction/repositories/SaleStatusRepository.java
@@ -3,5 +3,5 @@ package kattsyn.dev.ApartmentsUnderConstruction.repositories;
 import kattsyn.dev.ApartmentsUnderConstruction.entities.SaleStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SaleStatusRepository extends JpaRepository<SaleStatus, Long> {
+public interface SaleStatusRepository extends JpaRepository<SaleStatus, Integer> {
 }
